5/13
I made it so that when the user clicks the instructions button it provides a list of instructions for the game and from either the start screen or the instruction screen they can click space to start playing. Tomorrow I want to start work on the actual game and make the timer and point collection.