5/17
I made the lighting the right size and location and I made one of the food sprites that’s going to spawn randomly throughout but I haven’t made the random part. I’ve been trying to fix the opacity of the circle of light so that it actually looks like light but the set_alpha() isn’t working how it's supposed to, so I need to look into that next time.