6/7
I fixed the filter to make it darker so that it is harder to see food without the light, and I fixed the light so that it is actually a circle. I also added the points system and made it so when the player collides with the food it disappears and increases their points.