6/9
There were a few features I wanted to add over the weekend to make sure I’d have enough time to create my presentation, so I designed a jellyfish sprite and made it so when the player collides with the jellyfish it goes to a lose screen, I also implemented a timer but I need to test out exactly how long they should have to make sure the game is still possible. I also applied the point system that I created on Friday so the user can see the amount of food they need and watch it update as they get points, and if they get enough in the allotted time they win. I also made it so that the jellyfish and food would move on their own every few seconds to make them a bit harder to find.