6/10
I changed the light to be more white but also to have the effect of fading as it gets further away from the player but I’m not entirely sure if I’ll keep it, and I made the timer longer to make it easier to be able to win the game, and I added a second jellyfish. I also adjusted the size of the sprite for the player, because it was counting collisions that were within the rectangle but were blank space. I just have to fix one glitch where the jellyfish spawn in the same position as the player and the game immediately ends.