6/11
I spent a while messing around with the lighting trying to make it look better, but now I’m really happy with the results. I also fixed a few little errors, like the jellyfish moving off the screen, the text not being aligned in certain areas, ensuring all the sprites that need to be visible are, and adjusting the timer a bit more to ensure there is enough time to win.